third_party/WebKit/LayoutTests/http/tests/workers/worker-importScripts-onerror-sameorigin.html

<!DOCTYPE html>
<html>
<head>
    <script>
        window.jsTestIsAsync = true;
        window.isOnErrorTest = true;
    </script>
    <script src="/js-test-resources/js-test.js"></script>
</head>
<body>
    <script>
        description("Ensure that scripts imported into a Worker from same-origin hosts trigger detailed onerror messages.");

        var worker = new Worker('resources/worker-importscripts-onerror-sameorigin.js');

        var workerOnerror;
        worker.onmessage = function (e) {
            workerOnerror = e.data;
            shouldBeEqualToString("workerOnerror.message", "Uncaught This is a custom error message.");
            shouldBeEqualToString("workerOnerror.filename", "http://127.0.0.1:8000/workers/resources/worker-importScripts-throw.js");
            shouldEvaluateTo("workerOnerror.lineno", 1);
            shouldEvaluateTo("workerOnerror.colno", 1);
            shouldNotBe("workerOnerror.error", "null");
            shouldBeEqualToString("workerOnerror.error", "This is a custom error message.");
        };

        var onerrorMessage, onerrorURL, onerrorLine;
        worker.onerror = function (e) {
            pageOnerror = e;
            shouldBeEqualToString("pageOnerror.message", "Uncaught This is a custom error message.");
            shouldBeEqualToString("pageOnerror.filename", "http://127.0.0.1:8000/workers/resources/worker-importScripts-throw.js");
            shouldEvaluateTo("pageOnerror.lineno", 1);
            shouldEvaluateTo("pageOnerror.colno", 1);
            shouldBeNull("pageOnerror.error");

            e.preventDefault();
            finishJSTest();
        };
    </script>
</body>
</html>
